+++ src/main/java/io/github/natank25/epitechutils/module/EpitechConfigurable.kt
@@ -1,17 +1,12 @@
 package io.github.natank25.epitechutils.module
 
 import com.intellij.openapi.options.BoundConfigurable
-import com.intellij.openapi.options.Configurable
-import com.intellij.openapi.options.ConfigurationException
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.DialogPanel
-import com.intellij.openapi.util.NlsContexts.ConfigurableName
 import com.intellij.openapi.util.Weighted
 import com.intellij.ui.dsl.builder.bindText
 import com.intellij.ui.dsl.builder.panel
 import io.github.natank25.epitechutils.project.EpitechUtilsConfiguration
-import javax.swing.JComponent
-import javax.swing.JTextField
 
 class EpitechConfigurable internal constructor(private val project: Project) : BoundConfigurable(
     "Epitech"
@@ -25,14 +20,11 @@ class EpitechConfigurable internal constructor(private val project: Project) : B
         val settings = EpitechUtilsConfiguration.getInstance(project)
         return panel {
             row("Binary name :") {
-                textField().bindText(settings::BINARY_NAME).cellValidation {
-                    addInputRule("Binary name can not be empty.") {
-                        it.text.isEmpty()
-                    }
-                    addInputRule("Binary name can only contains a-z, A-Z, 0-9 _ . - characters.") {
-                        !it.text.matches(Regex("^[a-zA-Z0-9_.-]+\$"))
-                    }
-                }
+                textField().bindText(settings::BINARY_NAME)
+                    .addValidationRule("Binary name can not be empty")
+                        {field ->  field.text.isEmpty()}
+                    .addValidationRule("Binary name can only contains a-z, A-Z, 0-9 _ . - characters.")
+                        {field -> !field.text.matches(Regex("^[a-zA-Z0-9_.-]+\$"))}
             }
             row("Project name :"){
                 textField().bindText(settings::PROJECT_NAME)