+++ CHANGELOG.md
@@ -1,21 +1,41 @@
 # Epitech-Utils Changelog
 
 All notable changes to this project will be documented in this file.
-
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
 
 ## [Unreleased]
 
 ### Changed
+- Made the Makefile in the file template prettier.
+- Creates a default function when creating a C file.
+
+### Added
+- When creating a new Epitech Project, now also create a docker run configuration that generates the report for the Epitech Coding Style.
+- New file template for unit tests.
+- Shows most relevant file template first when creating a new file.
+- Now shows the content of the coding style report in a tool window.
+
+## [0.0.3] - 2024-12-01
+
+### Fixed
+
+- Plugin description should now not be duplicated on the Jetbrains Marketplace
+
+## [0.0.2] - 2024-12-01
+
+### Changed
+
 - Files are now created using an action and not a file template.
+- Entering a Github repository url now works and adds it as a remote.
+- When creating a project, automatically marks `include` and `src` folders as Source and Header files
+- When creating a new Epitech File, automatically reloads the Makefile project
 
 ## [0.0.1-beta.7] - 2024-11-23
 
 ### Changed
 
-- Plugin is now compatible with clion 23.3+ but not 23.2.3 
+- Plugin is now compatible with clion 23.3+ but not 23.2.3
 
 ## [0.0.1-beta.5]
 
@@ -35,8 +55,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
 
-[Unreleased]: https://github.com/Natank25/Epitech-Utils/compare/v0.0.1-beta.5...HEAD
+[Unreleased]: https://github.com/Natank25/Epitech-Utils/compare/v0.0.3...HEAD
+[0.0.3]: https://github.com/Natank25/Epitech-Utils/compare/v0.0.2...v0.0.3
+[0.0.2]: https://github.com/Natank25/Epitech-Utils/compare/v0.0.1-beta.7...v0.0.2
 [0.0.1-beta3]: https://github.com/Natank25/Epitech-Utils/commits/v0.0.1-beta3
+[0.0.1-beta.7]: https://github.com/Natank25/Epitech-Utils/compare/v0.0.1-beta.5...v0.0.1-beta.7
 [0.0.1-beta.5]: https://github.com/Natank25/Epitech-Utils/compare/v0.0.1-beta.3...v0.0.1-beta.5
 [0.0.1-beta.3]: https://github.com/Natank25/Epitech-Utils/compare/v0.0.1-beta.2...v0.0.1-beta.3
 [0.0.1-beta.2]: https://github.com/Natank25/Epitech-Utils/commits/v0.0.1-beta.2